# docs/apis/account-and-transaction/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Account and Transaction API",
    "description": "Account and Transaction API for Open Banking Connector Web App",
    "version": "web-app/5.1.0"
  },
  "paths": {
...
    "/aisp/balances": {
      "get": {
        "tags": [
          "Balances"
        ],
        "summary": "Read balances",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BalancesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalancesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalancesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/accounts/{externalApiAccountId}/balances": {
      "get": {
        "tags": [
          "Balances"
        ],
        "summary": "Read balances",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BalancesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalancesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalancesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/direct-debits": {
      "get": {
        "tags": [
          "Direct Debits"
        ],
        "summary": "Read direct debits",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DirectDebitsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectDebitsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectDebitsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/accounts/{externalApiAccountId}/direct-debits": {
      "get": {
        "tags": [
          "Direct Debits"
        ],
        "summary": "Read direct debits",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DirectDebitsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectDebitsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectDebitsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/party": {
      "get": {
        "tags": [
          "Parties"
        ],
        "summary": "Read party",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PartiesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartiesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartiesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/accounts/{externalApiAccountId}/party": {
      "get": {
        "tags": [
          "Parties"
        ],
        "summary": "Read party",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PartiesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartiesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartiesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/aisp/accounts/{externalApiAccountId}/parties": {
      "get": {
        "tags": [
          "Parties"
        ],
        "summary": "Read parties",
        "parameters": [
          {
            "name": "externalApiAccountId",
            "in": "path",
            "description": "External (bank) API ID of Account",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-obc-account-access-consent-id",
            "in": "header",
            "description": "ID of AccountAccessConsent used for request (obtained when creating consent)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-obc-modified-by",
            "in": "header",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Parties2Response"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parties2Response"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parties2Response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalancesResponse": {
        "type": "object",
        "properties": {
          "ExternalApiResponse": {
            "$ref": "#/components/schemas/OBReadBalance1"
          },
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to Balance read requests"
      },
      "DirectDebitsResponse": {
        "type": "object",
        "properties": {
          "ExternalApiResponse": {
            "$ref": "#/components/schemas/OBReadDirectDebit2"
          },
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to Direct Debit read requests"
      },
...
      "Links": {
        "type": "object",
        "properties": {
          "Self": {
            "type": "string",
            "description": "Gets the self.",
            "nullable": true
          },
          "First": {
            "type": "string",
            "description": "Gets the first.",
            "nullable": true
          },
          "Prev": {
            "type": "string",
            "description": "Gets the prev.",
            "nullable": true
          },
          "Next": {
            "type": "string",
            "description": "Gets the next.",
            "nullable": true
          },
          "Last": {
            "type": "string",
            "description": "Gets the last.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Links relevant to the payload."
      },
      "Meta": {
        "type": "object",
        "properties": {
          "TotalPages": {
            "type": "integer",
            "description": "Gets the total pages.",
            "format": "int32",
            "nullable": true
          },
          "FirstAvailableDateTime": {
            "type": "string",
            "description": "All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          },
          "LastAvailableDateTime": {
            "type": "string",
            "description": "All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Meta Data relevant to the payload."
      },
...
      "OBCreditDebitCode2Enum": {
        "enum": [
          "Credit",
          "Debit"
        ],
        "type": "string",
        "description": "Indicates whether the balance is a credit or a debit balance. \r\nUsage: A zero balance is considered to be a credit balance."
      },
...
      "OBParty2": {
        "type": "object",
        "properties": {
          "PartyId": {
            "type": "string",
            "description": "A unique and immutable identifier used to identify the customer resource. This identifier has no meaning to the account owner.",
            "nullable": true
          },
          "PartyNumber": {
            "type": "string",
            "description": "Number assigned by an agent to identify its customer.",
            "nullable": true
          },
          "PartyType": {
            "$ref": "#/components/schemas/OBExternalPartyType1CodeEnum"
          },
          "Name": {
            "type": "string",
            "description": "Name by which a party is known and which is usually used to identify that party.",
            "nullable": true
          },
          "FullLegalName": {
            "type": "string",
            "description": "Specifies a character string with a maximum length of 350 characters.",
            "nullable": true
          },
          "LegalStructure": {
            "type": "string",
            "description": "Legal standing of the party.",
            "nullable": true
          },
          "BeneficialOwnership": {
            "type": "boolean",
            "description": "Gets the beneficial ownership.",
            "nullable": true
          },
          "AccountRole": {
            "type": "string",
            "description": "A party’s role with respect to the related account.",
            "nullable": true
          },
          "EmailAddress": {
            "type": "string",
            "description": "Address for electronic mail (e-mail).",
            "nullable": true
          },
          "Phone": {
            "type": "string",
            "description": "Collection of information that identifies a phone number, as defined by telecom services.",
            "nullable": true
          },
          "Mobile": {
            "type": "string",
            "description": "Collection of information that identifies a mobile phone number, as defined by telecom services.",
            "nullable": true
          },
          "Relationships": {
            "$ref": "#/components/schemas/OBPartyRelationships1"
          },
          "Address": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OBParty2AddressItem"
            },
            "description": "Gets the address.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The OBParty2."
      },
...
      "OBReadBalance1": {
        "type": "object",
        "properties": {
          "Data": {
            "$ref": "#/components/schemas/OBReadBalance1Data"
          },
          "Links": {
            "$ref": "#/components/schemas/Links"
          },
          "Meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "description": "Response object OBReadBalance1 from UK Open Banking Read-Write Account and Transaction API spec\r\n<a href=\"https://github.com/OpenBankingUK/read-write-api-specs/blob/v3.1.10r5/dist/openapi/account-info-openapi.yaml\" />\r\nv3.1.10r5 <a />. Open Banking Connector will automatically\r\ntranslate <i>to</i> this from an older format for banks supporting an earlier spec version."
      },
      "OBReadBalance1Data": {
        "type": "object",
        "properties": {
          "Balance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OBReadBalance1DataBalanceItem"
            },
            "description": "Gets the balance.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The OBReadBalance1Data."
      },
...
      "OBReadBalance1DataBalanceItem": {
        "type": "object",
        "properties": {
          "AccountId": {
            "type": "string",
            "description": "A unique and immutable identifier used to identify the account resource. This identifier has no meaning to the account owner.",
            "nullable": true
          },
          "CreditDebitIndicator": {
            "$ref": "#/components/schemas/OBCreditDebitCode2Enum"
          },
          "Type": {
            "$ref": "#/components/schemas/OBBalanceType1CodeEnum"
          },
          "DateTime": {
            "type": "string",
            "description": "Indicates the date (and time) of the balance.All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time"
          },
          "Amount": {
            "$ref": "#/components/schemas/OBReadBalance1DataBalanceItemAmount"
          },
          "CreditLine": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OBReadBalance1DataBalancePropertiesItemsItem"
            },
            "description": "Gets the credit line.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Set of elements used to define the balance details."
      },
...
      "OBReadDirectDebit2": {
        "type": "object",
        "properties": {
          "Data": {
            "$ref": "#/components/schemas/OBReadDirectDebit2Data"
          },
          "Links": {
            "$ref": "#/components/schemas/Links"
          },
          "Meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "description": "The OBReadDirectDebit2."
      },
      "OBReadDirectDebit2Data": {
        "type": "object",
        "properties": {
          "DirectDebit": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OBReadDirectDebit2DataDirectDebitItem"
            },
            "description": "Gets the direct debit.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The OBReadDirectDebit2Data."
      },
      "OBReadDirectDebit2DataDirectDebitItem": {
        "type": "object",
        "properties": {
          "AccountId": {
            "type": "string",
            "description": "A unique and immutable identifier used to identify the account resource. This identifier has no meaning to the account owner.",
            "nullable": true
          },
          "DirectDebitId": {
            "type": "string",
            "description": "A unique and immutable identifier used to identify the direct debit resource. This identifier has no meaning to the account owner.",
            "nullable": true
          },
          "MandateIdentification": {
            "type": "string",
            "description": "Direct Debit reference. For AUDDIS service users provide Core Reference. For non AUDDIS service users provide Core reference if possible or last used reference.",
            "nullable": true
          },
          "DirectDebitStatusCode": {
            "$ref": "#/components/schemas/OBExternalDirectDebitStatus1CodeEnum"
          },
          "Name": {
            "type": "string",
            "description": "Name of Service User.",
            "nullable": true
          },
          "PreviousPaymentDateTime": {
            "type": "string",
            "description": "Date of most recent direct debit collection.All dates in the JSON payloads are represented in ISO 8601 date-time format. \r\nAll date-time fields in responses must include the timezone. An example is below:\r\n2017-04-05T10:43:07+00:00",
            "format": "date-time",
            "nullable": true
          },
          "Frequency": {
            "type": "string",
            "description": "Regularity with which direct debit instructions are to be created and processed.",
            "nullable": true
          },
          "PreviousPaymentAmount": {
            "$ref": "#/components/schemas/OBActiveOrHistoricCurrencyAndAmount0"
          }
        },
        "additionalProperties": false,
        "description": "Account to or from which a cash entry is made."
      },
      "OBReadParty2": {
        "type": "object",
        "properties": {
          "Data": {
            "$ref": "#/components/schemas/OBReadParty2Data"
          },
          "Links": {
            "$ref": "#/components/schemas/Links"
          },
          "Meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "description": "Response object OBReadParty2 from UK Open Banking Read-Write Account and Transaction API spec\r\n<a href=\"https://github.com/OpenBankingUK/read-write-api-specs/blob/v3.1.10r5/dist/openapi/account-info-openapi.yaml\" />\r\nv3.1.10r5 <a />. Open Banking Connector will automatically\r\ntranslate <i>to</i> this from an older format for banks supporting an earlier spec version."
      },
      "OBReadParty2Data": {
        "type": "object",
        "properties": {
          "Party": {
            "$ref": "#/components/schemas/OBParty2"
          }
        },
        "additionalProperties": false,
        "description": "The OBReadParty2Data."
      },
      "OBReadParty3": {
        "type": "object",
        "properties": {
          "Data": {
            "$ref": "#/components/schemas/OBReadParty3Data"
          },
          "Links": {
            "$ref": "#/components/schemas/Links"
          },
          "Meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "description": "Response object OBReadParty3 from UK Open Banking Read-Write Account and Transaction API spec\r\n<a href=\"https://github.com/OpenBankingUK/read-write-api-specs/blob/v3.1.10r5/dist/openapi/account-info-openapi.yaml\" />\r\nv3.1.10r5 <a />. Open Banking Connector will automatically\r\ntranslate <i>to</i> this from an older format for banks supporting an earlier spec version."
      },
      "OBReadParty3Data": {
        "type": "object",
        "properties": {
          "Party": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OBParty2"
            },
            "description": "Gets the party.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The OBReadParty3Data."
      },
...
      "Parties2Response": {
        "type": "object",
        "properties": {
          "ExternalApiResponse": {
            "$ref": "#/components/schemas/OBReadParty3"
          },
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to Parties read requests"
      },
      "PartiesResponse": {
        "type": "object",
        "properties": {
          "ExternalApiResponse": {
            "$ref": "#/components/schemas/OBReadParty2"
          },
          "Warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional list of warning messages from Open Banking Connector.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Response to Party read requests"
      },
...
    }
  }
}